requirements/base.txt
asgiref==3.2.3
Django==3.0.2
psycopg2-binary==2.8.4
django-tinymce==2.8.0
pytz==2019.3
sqlparse==0.3.0
celery==4.4.0
django-celery-beat==1.5.0
flower==0.9.3
django-environ==0.4.5
redis==3.3.11
django-allauth==0.41.0
django-braces==1.14.0